Cleanup of Communication service
