--- opencdx-communications/src/main/java/cdx/opencdx/communications/tasks/package-info.java
@@ -1,11 +1,0 @@
-/**
- * This is the scheduled tasks package project for the application.
- * <p>
- * Cron based tasks are scheduled here for execution.
- * </p>
- *
- * @author Safe Health LLC.
- * @version 1.0
- * @since 1.0
- */
-package cdx.opencdx.communications.tasks;
