+++ opencdx-communications/src/test/java/cdx/opencdx/communications/service/impl/OpenCDXEmailServiceImplTest.java
@@ -15,7 +15,6 @@
  */
 package cdx.opencdx.communications.service.impl;
 
-import cdx.opencdx.communications.service.OpenCDXEmailService;
 import cdx.opencdx.grpc.data.Attachment;
 import java.util.ArrayList;
 import java.util.List;
